MDEV-12746 rpl.rpl_parallel_optimistic_nobinlog fails committing            out of order at retry
The test failures were of two sorts. One is that the number of potential
temporary errors actually exceeds the default value of the slave retry
option.
Increased @@global.slave_transaction_retries fixes that part.

The 2nd issue was out of order commit by transactions that
are ordered after an erroring out transaction. The latter was possible
in the test thanks to the above retry failure.
Out-of-order committing was done despite the slave was stopping.

This is mended with refining conditions to retry. Specifically, a retrying
worker checks `rpl_parallel_entry::stop_on_error_sub_id` that
a failing predecessor could set to its own sub id.